+++ ydb/core/persqueue/pqtablet/common/ya.make
@@ -1,6 +1,7 @@
 LIBRARY()
 
 SRCS(
+    tracing_support.cpp
 )
 
 
@@ -13,8 +14,4 @@ PEERDIR(
 END()
 
 RECURSE_FOR_TESTS(
-#    ut
-#    dread_cache_service/ut
-#    ut/slow
-#    ut/ut_with_sdk
 )
