+++ ydb/core/persqueue/pqtablet/partition/partition.cpp
@@ -5,7 +5,7 @@
 #include "partition_common.h"
 #include "partition_compactification.h"
 #include <ydb/core/persqueue/pqtablet/common/logging.h>
-#include <ydb/core/persqueue/tracing_support.h>
+#include <ydb/core/persqueue/pqtablet/common/tracing_support.h>
 
 #include <ydb/library/wilson_ids/wilson.h>
 #include <ydb/core/base/appdata.h>
