+++ ydb/core/persqueue/pqtablet/pq_impl.cpp
@@ -5,7 +5,7 @@
 #include <ydb/core/persqueue/event_helpers.h>
 #include <ydb/core/persqueue/read.h>
 #include <ydb/core/persqueue/public/utils.h>
-#include <ydb/core/persqueue/tracing_support.h>
+#include <ydb/core/persqueue/pqtablet/common/tracing_support.h>
 #include <ydb/core/protos/config.pb.h>
 
 #include <ydb/core/base/tx_processing.h>
